+++ packages/core/global-navigation/examples/15-with-drawers-and-modal.js
@@ -224,7 +224,7 @@ export default class extends Component<{||}, NavState> {
           productNavigation={() => null}
           containerNavigation={() => null}
         >
-          <Fragment>
+          <div css={{ padding: '32px 40px' }}>
             <div>Page content</div>
             <button onClick={this.toggleCreateDrawer}>{`Enable ${
               this.state.isCreateDrawerEnabled
@@ -250,7 +250,7 @@ export default class extends Component<{||}, NavState> {
                 on closing the drawer
               </div>
             </div>
-          </Fragment>
+          </div>
         </LayoutManager>
       </NavigationProvider>
     );