+++ packages/core/checkbox/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/checkbox
 
+## 5.0.1
+- [patch] Checkbox now only fires onChange once [c78e59e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c78e59e)
+
 ## 5.0.0
 - [major] Checkbox refactored to remove the need for CheckboxStateless  [80e1925](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/80e1925)