+++ packages/growth/feature-flag-client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atlaskit/feature-flag-client
 
+## 1.0.2
+- [patch] MEP-103: Including flagKey in the attributes [afd42f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/afd42f7)
+
+## 1.0.1
+- [patch] MEP-103: Adding missing source to event shape [0e870d8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0e870d8)
+
+## 1.0.0
+- [major] Receive analyticsHandler instead of analyticClient [f082105](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f082105)
+
 ## 0.1.0
 - [minor] Implementing new version of the feature flag client [a7dbdbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a7dbdbb)