+++ packages/search/global-search/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/global-search
 
+## 5.7.2
+- [patch] fix jira response parser [1502daa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1502daa)
+
 ## 5.7.1
 - [patch] remove forward slash that breaks jira requests [08d26e0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/08d26e0)